requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.1
Pillow==10.2.0
python-magic==0.4.27
numpy==1.26.4
streamlit-drawable-canvas==0.9.3
